refactor(chat): extract shared chat transport utilities
